requirements.txt
asgiref
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.10
coverage==5.5
Django==3.2.5
django-countries==7.2.1
flake8==3.9.2
flake8-django==1.1.2
idna==3.3
isort==5.9.3
mccabe==0.6.1
Pillow==8.3.1
psycopg2-binary
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2020.4
pyuca==1.2
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse
testfixtures==6.18.3
toml==0.10.2
typing-extensions
urllib3==1.26.8
